+++ utils/vis_tool.py
@@ -62,7 +62,7 @@ def vis_image(img, ax=None):
         ax = fig.add_subplot(1, 1, 1)
     # CHW -> HWC
     img = img.transpose((1, 2, 0))
-    ax.imshow(img.astype(np.uint8))
+    ax.imshow(img.astype(np.uint32))
     return ax
